+++ oracle_common/core/config.go
@@ -38,7 +38,7 @@ type CardanoChainConfigUtxo struct {
 
 type ColoredCoinEvm struct {
 	TokenName       string `json:"tokenName"`
-	ColoredCoinID   uint8  `json:"coloredCoinID"`
+	ColoredCoinID   uint16 `json:"coloredCoinID"`
 	ContractAddress string `json:"contractAddress"`
 }
